+++ source/components/panel/panel.js
@@ -1,4 +1,4 @@
-/* global Metro, METRO_ANIMATION_DURATION */
+/* global Metro, METRO_ANIMATION_DURATION, Cake */
 (function(Metro, $) {
     'use strict';
     var Utils = Metro.utils;
@@ -90,7 +90,7 @@
 
                 if (item.attr && typeof item.attr === 'object') {
                     $.each(item.attr, function(k, v){
-                        customButton.attr($.dashedName(k), v);
+                        customButton.attr(Cake.dashedName(k), v);
                     });
                 }